# swagger.yaml
swagger: "2.0"
info:
  description: |
    Backend API to support HotelHopper.cf \
    To see the standardized api responses see the models at the bottom\: 
    SuccessfulResponse for 200, BadRequestResponse for 400, and UnauthorizedResponse for 401 \
    More HTTP Status Codes may be added in the future and similar models will be made avialable.
  version: "1.0.0"
  title: "Hotel Hopper"
host: "staging-api.hotelhopper.cf"
...
schemes:
- "https"
paths:
...
  /hotels:
    get:
      tags:
      - "hotels"
      summary: "Returns an array of hotels"
      description: "Returns an array of hotels based on the given searh parameters. Pagination can be acheived using the `page` and `perPage` parameters. By default `page` is **1** and `perPage` is **15**. Only hotels which have enough rooms to acommodate the number of `persons` will be returned. Hotels will also be restricted a 5 mile radius from the given coordinates."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
      - name: latitude
        in: query
        type: number
        required: true
        description: 'Provide a float value for the latitude of the city center being searched. Ex: 37.3440232'
      - name: longitude
        in: query
        type: number
        required: true
        description: 'Provide a float value for the longitude of the city center being searched. Ex: -121.8738311'
      - name: startDate
        in: query
        type: string
        required: true
        description: 'Must be a date in the format YYYY-MM-DD'
      - name: endDate
        in: query
        type: string
        required: true
        description: 'Must be a date in the format YYYY-MM-DD'
      - name: persons
        in: query
        type: integer
        required: true
        description: 'The amount of people traveling'
      - name: page
        in: query
        type: integer
        description: 'Skips (page * perPage) number of hotels'
        default: 1
      - name: perPage
        in: query
        type: integer
        description: 'The number of hotels to be returned'
        default: 15
      responses:
        200:
          description: "OK"
          schema:
            $ref: '#/definitions/HotelSearchResponse'
        400:
          description: "Bad Request"
          schema:
            $ref: '#/definitions/BadRequestResponse'
  /hotels/:id:
    get:
      tags:
      - "hotels"
      summary: "Returns a hotels details, available rooms, and amenities"
      description: "Returns the full details of the hotel and the available rooms given the `startDate` and `endDate` of the trip."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
      - name: startDate
        in: query
        type: string
        required: true
        description: 'Must be a date in the format YYYY-MM-DD'
      - name: endDate
        in: query
        type: string
        required: true
        description: 'Must be a date in the format YYYY-MM-DD'
      responses:
        200:
          description: "OK"
          schema:
            $ref: '#/definitions/HotelInfoResponse'
        400:
          description: "Bad Request"
          schema:
            $ref: '#/definitions/BadRequestResponse'
        401:
          description: "Unauthorized"
          schema:
            $ref: '#/definitions/UnauthorizedResponse'

  # Reservations
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  BadRequestResponse:
    type: object
    properties:
      error:
        type: boolean
      message:
        type: string
        description: "The first found validation error"
      validationErrors:
        type: object
        properties:
          <name of parameter>:
            type: array
            description: "Array of the problems with the parameter value"
            items:
              type: string
    required:
    - error
    - message
  UnauthorizedResponse:
    type: object
    properties:
      error:
        type: boolean
      message:
        type: string
        enum:
          - "No user found with JWT"
          - "Invalid JWT format"
          - "Expired JWT"
    required:
    - error
    - message
  HotelSearchResponse:
    type: object
    properties:
      error:
        type: boolean
      data:
        type: array
        items:
          $ref: '#/definitions/Hotel'
  HotelInfoResponse:
    type: object
    properties:
      error:
        type: boolean
      data:
        $ref: '#/definitions/FullHotel'
...
  Hotel:
    type: object
    properties:
      hotelId:
        type: integer
      title: 
        type: string
      description:
        type: string
      street:
        type: string
      city:
        type: string
      state:
        type: string
      zip:
        type: string
      country:
        type: string
      address:
        type: string
      mapURL:
        type: string
      imageURL:
        type: string
      stars:
        type: integer
      rating:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      lowestPrice:
        type: number
  FullHotel:
    type: object
    properties:
      hotelId:
        type: integer
      title: 
        type: string
      description:
        type: string
      street:
        type: string
      city:
        type: string
      state:
        type: string
      zip:
        type: string
      country:
        type: string
      address:
        type: string
      mapURL:
        type: string
      imageURL:
        type: string
      stars:
        type: integer
      rating:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      rooms:
        type: array
        items:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
            persons:
              type: integer
            beds:
              type: integer
            available:
              type: integer
      hotelAmenities:
        type: array
        items:
          type: object
          properties:
            amenityId:
              type: integer
            amenity:
              type: object
              properties:
                title:
                  type: string
                imageUrl:
                  type: string
...